+++ packages/crayons-extended/custom-objects/src/components/form-builder/utils/form-builder-utils.ts
@@ -264,3 +264,14 @@ export function createUUID() {
   );
   return uuid;
 }
+
+export function checkIfCustomToggleField(
+  productName = 'CUSTOM_OBJECTS',
+  fieldName
+) {
+  const dbConfig = formMapper[productName];
+  return (
+    dbConfig?.config?.boolShowCustomToggle &&
+    fieldName === dbConfig?.config?.showCustomToggleField
+  );
+}